requirements.txt
aiohttp==3.11.10
cachetools==5.3.0
httpx==0.28.1
python-dotenv==1.0.1
python-telegram-bot==21.9
qrcode==8.0
Requests==2.32.3
solana==0.36.1
solders==0.23.0
supabase==2.10.0
